refactor: simplify Slack launcher test seam
